requirements.txt
rich~=13.7.1
chardet~=5.2.0
jieba~=0.42.1
tqdm~=4.66.5
Cython~=3.0.11
Pillow~=10.4.0
future~=1.0.0
questionary~=2.0.1
argparse
faker
pyperclip~=1.9.0
zxcvbn~=4.4.28
fastapi~=0.112.0
uvicorn~=0.30.6
pymupdf==1.19.6
win10toast
matplotlib
psutil
requests~=2.32.3
nuitka
PyQt5~=5.15.11
opencv-python~=4.10.0.84
pydantic~=2.8.2
reportlab~=4.2.2
numpy~=2.0.1
scikit-image~=0.24.0
fpdf~=1.7.2
pandas~=2.2.2
uvicorn>=0.30.6
fastapi>=0.112.1
PyYAML>=6.0.2
requests>=2.32.3
PyExecJS>=1.5.1
pydantic>=2.8.2
starlette~=0.38.2